Fixed Linux, MacOS, Windows warnings
* Started refactoring

* Test

* Added second Workflow

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Fixed tests

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test wheels

* Test

* Test

* Test

* Test

* Test

* Test Warnings

* Test

* Test Cmake 3.8 and find python

* Test

* Test

* Test

* Test

* Test

* Test

* Fixed conan override outputs

* Fixed python module generation

* Test

* Test

* Cleaned workflows and cmakelist

* Disable build docs by default

* Added resources to python module + Updated python docs

* Updated GameRenderer section + Updated mains loading correct default config

* Updated installation docs

* Disabled workflow wheels on push

* W1 level fixed warnings

* W2 level fixed warnings

* W3 level fixed warnings + render error

* Test warnings level w3

* Test warnings level w1

* Test

* Test

* Test

* Testing

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* FIxed last windows errors w3

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Test

* Cleaned code + Test wheels pybind enabled

* Python wheel test

* Test

* Test

* Test

* Test

* Test

* Revert "Test"

This reverts commit 2022a43fc2dc4cc9188219d1a4ce9abd8fbe1086.

+++ src/stratega/include/Stratega/Utils/IteratorToPtrVector.h
@@ -63,7 +63,7 @@ namespace SGA
 		static std::vector<BaseType*> convert(Iterator begin, Iterator end)
 		{
 			std::vector<BaseType*> vec;
-			vec.reserve(std::distance(begin, end));
+			vec.reserve(static_cast<size_t>(std::distance(begin, end)));
 			for (auto it = begin; it != end; ++it)
 			{
 				vec.emplace_back(*it ? it->get() : nullptr);
